refactor: restructure project directory and update import paths
