+++ app/resources/icons/reset.png
