+++ app/resources/icons/rotate_left.png
