+++ app/resources/icons/rotate_right.png
