+++ app/screens/utils/custom_logging.py
