+++ app/screens/utils/tensorboard_utils.py
@@ -5,8 +5,8 @@ import webbrowser
 
 from tensorboard import program
 
-from screens.custom_logging import get_logger
-from screens.db import DB
+from app.screens.utils.custom_logging import get_logger
+from app.screens.utils.db import DB
 
 logger = get_logger(__name__)
