+++ app/screens/view/image_screen.py
@@ -15,9 +15,9 @@ from kivy.uix.textinput import TextInput
 from kivymd.uix.floatlayout import MDFloatLayout
 from kivymd.uix.selectioncontrol import MDCheckbox
 
-from screens.additional import BaseScreen, ImageMDButton, MDLabelBtn
-from screens.custom_logging import get_logger
-from utils import extend_key
+from app.screens.utils.additional import BaseScreen, ImageMDButton, MDLabelBtn
+from app.screens.utils.custom_logging import get_logger
+from app.screens.utils.utils import extend_key
 
 logger = get_logger(__name__)
