+++ scripts/install/install_torch_lin_cu124.sh
